+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The following table shows which versions of MapReader are compatible with which 
 ## Pre-release
 _Add new changes here_
 
+## [v1.8.3](https://github.com/Living-with-machines/MapReader/releases/tag/v1.8.3) (20256-09-13)
+
 ### Fixed
 
 - Fixes the `model_summary` method in the `ClassifierContainer` class ([#574](https://github.com/maps-as-data/MapReader/pull/574))